pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wealthsense-ai"
version = "0.1.0"
description = "Deep-learning financial advisor with forecasting and goal planning"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.5.0",
  "torch>=2.3.0",
  "yfinance>=0.2.54",
  "streamlit>=1.44.0",
  "plotly>=5.24.0",
  "python-dotenv>=1.0.1",
  "anthropic>=0.49.0",
  "fastapi>=0.128.0",
  "uvicorn>=0.39.0",
  "requests>=2.32.0",
  "supabase>=2.28.0",
  "pytest>=8.4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
